Reapply "[INJIMOB-3702] Move private QR function into helper"
This reverts commit 7c07e266ed86ffc3edfbac302342882d2aeaffa7.
