+++ Sources/pixelpass/pixelpass.swift
@@ -107,33 +107,12 @@ public class PixelPass {
         return base45EncodedString
     }
     
-    private func convertQRDataIntoQRImage(
-        qrText: String,
-        ecc: ECC = .L
-    ) -> Data? {
-
-        let data = qrText.data(using: .ascii)
-
-        if let filter = CIFilter(name: "CIQRCodeGenerator") {
-            filter.setValue(data, forKey: "inputMessage")
-            filter.setValue(ecc.rawValue, forKey: "inputCorrectionLevel")
-
-            if let qrImage = filter.outputImage {
-                let context = CIContext()
-                if let cgImage = context.createCGImage(qrImage, from: qrImage.extent) {
-                    return UIImage(cgImage: cgImage).pngData()
-                }
-            }
-        }
-        return nil
-    }
-    
     public func generateQRCodeFromQRData(
         qrData: String,
         ecc: ECC = .L
     ) -> Data? {
 
-        return convertQRDataIntoQRImage(
+        return QRCodeHelper.generateQRImage(
             qrText: qrData,
             ecc: ecc
         )
@@ -145,7 +124,7 @@ public class PixelPass {
                 return nil
             }
 
-            return convertQRDataIntoQRImage(
+        return QRCodeHelper.generateQRImage(
                 qrText: qrText + header,
                 ecc: ecc
             )
